push_swap

+++ push_swap/libft/ft_atoi.c
@@ -18,7 +18,7 @@ static	int		ft_isspace(char c)
 			c == ' ');
 }
 
-int				ft_atoi(const char *str)
+long				ft_atoi(const char *str)
 {
 	int		i;
 	long	num;
